+++ game/client.py
@@ -8,7 +8,7 @@ def main():
 def play(chess):
     try:
         # print(chess.show_board())
-        print('Turno: ', chess.turn  )
+        print(f'Turno: {chess.turn}')
         from_row = int(input('Fila Inicial: '))
         from_col = int(input('Columna Inicial: '))
         to_row = int(input('Fila Final: '))